ID type fix
Valamiért Integerek voltak az ID-ink, míg a backenden (jól) Long típusúak. Most már nálunk is Long.

+++ UntitledTemalab/app/src/main/java/hu/bme/aut/untitledtemalab/data/JobData.kt
@@ -1,13 +1,13 @@
 package hu.bme.aut.untitledtemalab.data
 
 //Mark as "data class"?
-class JobData (val jobId: Int,
+class JobData (val jobId: Long,
                val jobName: String,
                val packageSize: PackageSize,
                val payment: Int,
                val jobIssuedDate: String,
                val deadline: String,
-               val ownerID: Int){
+               val ownerID: Long){
     //Tagváltozók
     var status: JobStatus = JobStatus.Pending
     // TODO: Ezt a senderRating-et miért tettem ide? ez a hozzá tartozó User-ben van
