+++ UntitledTemalab/app/src/main/java/hu/bme/aut/untitledtemalab/features/joblistfeatures/currentjobs/CurrentJobsContainerFragment.kt
@@ -14,7 +14,7 @@ import kotlin.properties.Delegates
 
 class CurrentJobsContainerFragment : Fragment() {
 
-    private var userId by Delegates.notNull<Int>()
+    private var userId by Delegates.notNull<Long>()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
