+++ UntitledTemalab/app/src/main/java/hu/bme/aut/untitledtemalab/features/joblistfeatures/jobboard/JobBoardContainerRepository.kt
@@ -3,7 +3,7 @@ package hu.bme.aut.untitledtemalab.features.joblistfeatures.jobboard
 import hu.bme.aut.untitledtemalab.network.NetworkManager
 import hu.bme.aut.untitledtemalab.network.response.UserDataResponse
 
-class JobBoardContainerRepository(private val userId: Int){
+class JobBoardContainerRepository(private val userId: Long){
 
     suspend fun getUsersCargoDetails(): UserDataResponse{
         return try{
